Hack up a -devel package for spdk
The spdk wants to use the fio source to build rather than a
more traditional -devel package (which fio isn't yet providing),
so for the time being, add the files needed to the debuginfo
package and spdk can use that.

This will need to be corrected more properly in the future.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>